requirements.txt
playwright>=1.40.0
opencv-python>=4.8.0
numpy>=1.24.0
PyYAML>=6.0.0
scikit-image>=0.22.0
torch>=2.6.0
torchvision>=0.21.0
